Module Tezos_client_sapling_009_PsFLoren.Wallet

module Mnemonic : sig ... end

Mnemonic of 24 common english words from which a key can be derived. The mnemonic follows the BIP-39 spec.

val register : Tezos_client_base.Client_context.#full -> ?⁠force:bool -> ?⁠unencrypted:bool -> Tezos_client_base.Bip39.t -> string -> Tezos_sapling.Core.Client.Viewing_key.t Tezos_base__TzPervasives.tzresult Lwt.t

Add to the wallet a new spending key derived from a mnemonic and identified by an alias. The wallet is updated and the corresponding viewing key is returned. If force it will overwrite an existing alias.

val derive : Tezos_client_base.Client_context.#full -> ?⁠force:bool -> ?⁠unencrypted:bool -> string -> string -> int -> (string * Tezos_sapling.Core.Client.Viewing_key.t) Tezos_base__TzPervasives.tzresult Lwt.t

derive parent child index derives a key with alias child from an existing key with alias parent at index using ZIP32. If a new index is required the state of the wallet is updated. The path and viewing key corresponding to the generated key are returned.

val find_vk : Tezos_client_base.Client_context.#full -> string -> Tezos_sapling.Core.Client.Viewing_key.t Tezos_base__TzPervasives.tzresult Lwt.t

val new_address : Tezos_client_base.Client_context.#full -> string -> int option -> (Tezos_sapling.Core.Client.Spending_key.t * Tezos_sapling.Core.Client.Viewing_key.index * Tezos_sapling.Core.Client.Viewing_key.address) Tezos_base__TzPervasives.tzresult Lwt.t

Generate a new address. If an optional index is provided, try to derive the address at this index, otherwise use the first viable one. Not all indexes correspond to a valid address so successive ones are tried. Once a valid index is found it is recorded in the wallet. Return also the corresponding sk and vk to avoid asking the user multiple times for the description password.

val export_vk : Tezos_client_base.Client_context.#full -> string -> Tezos_base__TzPervasives.Data_encoding.Json.json Tezos_base__TzPervasives.tzresult Lwt.t